requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-allauth==0.41.0
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.0
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.6
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==5.0.1
omdb==0.10.1
